+++ Bugsnag/KSCrash/Source/KSCrash/Recording/Sentry/BSG_KSCrashSentry_CPPException.mm
@@ -24,11 +24,12 @@
 
 #import <Foundation/Foundation.h>
 
+#include "BSGDefines.h"
+#include "BSG_KSCrashC.h"
 #include "BSG_KSCrashSentry_CPPException.h"
 #include "BSG_KSCrashSentry_Private.h"
+#include "BSG_KSCrashStringConversion.h"
 #include "BSG_KSMach.h"
-#include "BSG_KSCrashC.h"
-#include "BSGDefines.h"
 
 //#define BSG_KSLogger_LocalLevel TRACE
 #include "BSG_KSLogger.h"
@@ -106,7 +107,6 @@ void __cxa_throw(void *thrown_exception, std::type_info *tinfo,
 static void CPPExceptionTerminate(void) {
     BSG_KSLOG_DEBUG("Trapped c++ exception");
 
-    bool isNSException = false;
     char descriptionBuff[DESCRIPTION_BUFFER_LENGTH];
     const char *name = NULL;
     const char *crashReason = NULL;
@@ -132,15 +132,13 @@ static void CPPExceptionTerminate(void) {
     try {
         throw;
     } catch (NSException *exception) {
-        BSG_KSLOG_DEBUG("Detected NSException. Recording details and letting "
-                        "the current NSException handler deal with it.");
-        isNSException = true;
-        // recordException() doesn't call beginHandlingCrash()
-        if (bsg_kscrashsentry_isNSExceptionHandlerInstalled() &&
-            bsg_kscrashsentry_beginHandlingCrash(bsg_ksmachthread_self())) {
-            bsg_recordException(exception);
-            bsg_kscrashsentry_endHandlingCrash();
+        if (bsg_g_originalTerminateHandler != NULL) {
+            BSG_KSLOG_DEBUG("Detected NSException. Passing to the current NSException handler.");
+            bsg_g_originalTerminateHandler();
+        } else {
+            BSG_KSLOG_DEBUG("Detected NSException, but there was no original C++ terminate handler.");
         }
+        return;
     } catch (std::exception &exc) {
         strlcpy(descriptionBuff, exc.what(), sizeof(descriptionBuff));
         crashReason = descriptionBuff;
@@ -148,34 +146,49 @@ static void CPPExceptionTerminate(void) {
         strlcpy(descriptionBuff, exc->what(), sizeof(descriptionBuff));
         crashReason = descriptionBuff;
     }
-#define CATCH_VALUE(TYPE, PRINTFTYPE)                                          \
-    catch (TYPE value) {                                                       \
-        snprintf(descriptionBuff, sizeof(descriptionBuff), "%" #PRINTFTYPE,    \
-                 value);                                                       \
-        crashReason = descriptionBuff;                                         \
+#define CATCH_INT(TYPE)                                           \
+    catch (TYPE value) {                                          \
+        bsg_int64_to_string(value, descriptionBuff);              \
+        crashReason = descriptionBuff;                            \
     }
-    CATCH_VALUE(char, d)
-    CATCH_VALUE(short, d)
-    CATCH_VALUE(int, d)
-    CATCH_VALUE(long, ld)
-    CATCH_VALUE(long long, lld)
-    CATCH_VALUE(unsigned char, u)
-    CATCH_VALUE(unsigned short, u)
-    CATCH_VALUE(unsigned int, u)
-    CATCH_VALUE(unsigned long, lu)
-    CATCH_VALUE(unsigned long long, llu)
-    CATCH_VALUE(float, f)
-    CATCH_VALUE(double, f)
-    CATCH_VALUE(long double, Lf)
-    CATCH_VALUE(char *, s)
+#define CATCH_UINT(TYPE)                                          \
+    catch (TYPE value) {                                          \
+        bsg_uint64_to_string(value, descriptionBuff);             \
+        crashReason = descriptionBuff;                            \
+    }
+#define CATCH_DOUBLE(TYPE)                                        \
+    catch (TYPE value) {                                          \
+        bsg_double_to_string((double)value, descriptionBuff, 16); \
+        crashReason = descriptionBuff;                            \
+    }
+#define CATCH_STRING(TYPE)                                        \
+    catch (TYPE value) {                                          \
+        strncpy(descriptionBuff, value, sizeof(descriptionBuff)); \
+        descriptionBuff[sizeof(descriptionBuff)-1] = 0;           \
+        crashReason = descriptionBuff;                            \
+    }
+
+    CATCH_INT(char)
+    CATCH_INT(short)
+    CATCH_INT(int)
+    CATCH_INT(long)
+    CATCH_INT(long long)
+    CATCH_UINT(unsigned char)
+    CATCH_UINT(unsigned short)
+    CATCH_UINT(unsigned int)
+    CATCH_UINT(unsigned long)
+    CATCH_UINT(unsigned long long)
+    CATCH_DOUBLE(float)
+    CATCH_DOUBLE(double)
+    CATCH_DOUBLE(long double)
+    CATCH_STRING(char *)
     catch (...) {
     }
 
 after_rethrow:
     bsg_g_captureNextStackTrace = (bsg_g_installed != 0);
 
-    if (!isNSException &&
-        bsg_kscrashsentry_beginHandlingCrash(bsg_ksmachthread_self())) {
+    if (bsg_kscrashsentry_beginHandlingCrash(bsg_ksmachthread_self())) {
 
 #if BSG_HAVE_MACH_THREADS
         BSG_KSLOG_DEBUG("Suspending all threads.");