+++ Tests/BugsnagTests/BugsnagConfigurationTests.m
@@ -872,6 +872,8 @@ NSString * const kBugsnagUserUserId = @"BugsnagUserUserId";
 #if !TARGET_OS_WATCH
     [config setSendThreads:BSGThreadSendPolicyUnhandledOnly];
 #endif
+    [config addPlugin:(id)[NSNull null]];
+
     BugsnagOnSendErrorBlock onSendBlock1 = ^BOOL(BugsnagEvent * _Nonnull event) { return true; };
     BugsnagOnSendErrorBlock onSendBlock2 = ^BOOL(BugsnagEvent * _Nonnull event) { return true; };
 
@@ -922,6 +924,10 @@ NSString * const kBugsnagUserUserId = @"BugsnagUserUserId";
     XCTAssertEqualObjects(clone.notifier.name, config.notifier.name);
     XCTAssertEqualObjects(clone.notifier.version, config.notifier.version);
     XCTAssertEqualObjects(clone.notifier.url, config.notifier.url);
+
+    // Plugins
+    XCTAssert([clone.plugins containsObject:[NSNull null]]);
+    XCTAssertNoThrow([clone.plugins removeObject:[NSNull null]]);
 }
 
 - (void)testMetadataMutability {