+++ features/fixtures/shared/scenarios/Scenario.m
@@ -6,8 +6,6 @@
 
 #import <objc/runtime.h>
 
-extern void bsg_kscrash_setPrintTraceToStdout(bool printTraceToStdout);
-
 extern bool bsg_kslog_setLogFilename(const char *filename, bool overwrite);
 
 extern void bsg_i_kslog_logCBasic(const char *fmt, ...) __printflike(1, 2);
@@ -94,8 +92,6 @@ static char ksLogPath[PATH_MAX];
 
 - (void)startBugsnag {
     [Bugsnag startWithConfiguration:self.config];
-
-    bsg_kscrash_setPrintTraceToStdout(true);
 }
 
 - (void)didEnterBackgroundNotification {